+++ lib/dia/sandbox.rb
@@ -1,11 +1,8 @@
 module Dia
   
-  class SandBox
+  class Sandbox
   
     include Dia::CommonAPI
-    
-    # We're going to deprecate Dia::SandBox in favor of Dia::Sandbox in the next release.
-    Dia::Sandbox = SandBox
     
     attr_accessor :app_path
     attr_accessor :profile